+++ priv/repo/migrations/20191104143558_fix_null_muted_notification_fields.exs
@@ -1,0 +1,7 @@
+defmodule Pleroma.Repo.Migrations.FixNullMutedNotificationFields do
+  use Ecto.Migration
+
+  def change do
+    execute("update users set info = safe_jsonb_set(info, '{muted_notifications}', '[]'::jsonb, true) where local = true and info->'muted_notifications' = 'null'::jsonb")
+  end
+end